Actions de login e signup acopladas e signup redireciona para feed logado
